[ci skip] WIP - Change environment type
The environment needs access to the ids and tyvars of inserted
local types and expressions to create the substitution. This means
that VarEnv is not suitable, as it uses only the unique as a key.

The new implementation seems to work, but we should update
Clash.Driver.Utils and see if we can get the global and prim heaps
back after running the partial evaluator.