[ci skip] WIP - Apply to Data and Prim in apply / applyTy
The apply / applyTy functions were not applying to data constructors
and prims. This has been amended, and the now unnecessary
evalApplication function has been removed.

There is a problem with some tests failing because they attempt
to apply the State# RealWorld to an IO or IO newtype, resulting in
one argument too many according to the evalautor.