Lower timeout for CI jobs
Most CI jobs do not need a timeout as long as they currently have,
as the long timeout is only for cases when there is no cache for
nix. To prevent the CI holding onto runaway jobs for a long time,
the timeout for most tasks is lowered to 2 hours.